backend/requirements.txt
asgiref==3.11.0
dj-database-url==3.0.1
Django==4.2.7
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==23.0.0
joblib==1.3.2
numpy==1.26.2
packaging==25.0
pandas==2.1.3
Pillow==10.1.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
scikit-learn==1.3.2
scipy==1.16.3
six==1.17.0
sqlparse==0.5.4
threadpoolctl==3.6.0
tzdata==2025.2
whitenoise==6.11.0
